Made more stuff async
