+++ frontend/apps/crates/entry/admin/src/locale/components/mod.rs
@@ -1,3 +1,4 @@
+pub(super) mod locale_outer;
 pub(super) mod table;
 pub(super) mod entry_row;
 pub(super) mod select_columns;
